Support Flask application alongside the Django app
There's a new 'flask' role that's set up to pull a Flask application,
to be served alongside the Django application in production.

In MITOC's specific case, this is the membership API, which occupies the
URLs '/members/*' - everything else still directs to Django.

This new role is very similar to the Django configuration (it's run by
Gunicorn, has its own user, etc.).

Some notable differences in the flask role (vs web role):
- We use Python 3.6
- No maintenance page needed (It's intended to always be up)
- libssl, postgres, and a bunch of other env vars not needed
- No need for Google credentialss

The 'setup Flask' task is also much, much smaller.