Also create MySQL database locally
When initializing the app with Django 1.11, `migrate` attempts to make a
connection with all the defined connections. As a result, this raises an
exception when running the Vagrant playbook, since the `geardb`
connection doesn't actually connect to anything.

To correct this, also create a MySQL database locally and configure it
for use as the `geardb`.