Use new DocuSign certificates
DocuSign certificates are being renewed for 2018, so completed envelope
events will no longer be signed with the current certificate.
Accordingly, our NGINX configuration would reject valid notifications
from the production DocuSign service.

We need to swap over the expected fingerprints to support both the new
certificate and the current certificate. This way, there's no
interruption in API service when the actual switchover takes place.

In production, the DocuSign user is not an admin, so I have no control
over when we actually switch to new certificates. On April 18th, admins
will have the ability to swap over to the new certificate. On May 23rd,
the new certificate will be used no matter what. Depending on what IS&T
wishes to do, I can expect a switchover as soon as April 18, or as late
as May 23rd.