Require that all images be served over HTTPS
I manually corrected the few trips that were hotlinking HTTP images. In
most cases, the host now supports HTTPS. In others, it was appropriate
to just remove the image entirely since it wasn't clear that the poster
had permission to use the image.

This will remove all warnings about mixed content on a page that should
be all HTTPS.