Describe new redaction system
Previously, I'd been removing the entirety of `production.yml` from the
Git history on the public version of this repository. That's not
strictly necessary, since there are a lot of non-secrets included in
`production.yml`. Use a new system that redacts secrets but leaves other
text & comments in that file.