Change PRNG secret value
I accidentally included the production PRNG seed in a commit. This is
not a big deal from a security perspective - the only value this secret
provides is letting a participant know lottery ranking in advance,
assuming that the lottery is run at a precise time & date with an exact
set of participating users, with an exact set of IDs, etc.

In other words, the scope of being able to potentially exploit this
secret for personal gain is extremely limited.

So long as this secret is changed before an actual lottery is run with
it, there's zero harm in it being out in plaintext.