Update IP address for new instance
We recently migrated to a `t2.medium` (paid for on-demand). A `t2.micro`
is plenty of muscle for the day-to-day needs, but we've seen some
critical load during WS lectures. Upgrading to a medium for now
addresses some key needs:

- A robust webserver for during WS
- Ability to run `npm build` directly on the server for the new FE
- Ability to run `collectstatic` directly on the host for the legacy FE

In the future, we can go right back to `t2.micro`, assuming that we've
solved the above problems. Better solutions:

1. (For WS load): Have an auto-scaling pool of webserver workers, behind
   an ELB. Possibly just containerize a worker for really fast spin-up.
2. Change the playbook to do collection of static files directly on the
   caller, then SCP over to the webserver (also easily accomplished with
   containerization)